+++ TktkDirectX12GameLib/TktkDX12GameLib/inc/TktkDX12Game/DXGameResource/DXGameShaderResouse/MeshResouse/MeshResource.h
@@ -65,8 +65,8 @@ namespace tktk
 			size_t skeletonHandle,
 			size_t curMotionHandle,
 			size_t preMotionHandle,
-			size_t curFrame,
-			size_t preFrame,
+			float curFrame,
+			float preFrame,
 			float amount
 		);
