+++ TktkDirectX12GameLib/TktkDX12GameLib/inc/TktkDX12Game/_MainManager/DX12GameManager.h
@@ -570,8 +570,8 @@ namespace tktk
 			size_t skeletonHandle,
 			size_t curMotionHandle,
 			size_t preMotionHnadle,
-			size_t curFrame,
-			size_t preFrame,
+			float curFrame,
+			float preFrame,
 			float amount
 		);
