+++ TktkDirectX12GameLib/TktkDX12GameLib/src/TktkDX12Game/DXGameResource/DXGameResource.cpp
@@ -322,7 +322,7 @@ namespace tktk
 		return m_dxGameShaderResouse->getMotionEndFrameNo(handle);
 	}
 
-	void DXGameResource::updateMotion(size_t skeletonHandle, size_t curMotionHandle, size_t preMotionHandle, size_t curFrame, size_t preFrame, float amount)
+	void DXGameResource::updateMotion(size_t skeletonHandle, size_t curMotionHandle, size_t preMotionHandle, float curFrame, float preFrame, float amount)
 	{
 		m_dxGameShaderResouse->updateMotion(skeletonHandle, curMotionHandle, preMotionHandle, curFrame, preFrame, amount);
 	}
