+++ TktkDirectX12GameLib/TktkDX12GameLib/src/TktkDX12Game/DXGameResource/DXGameShaderResouse/MeshResouse/MeshResource.cpp
@@ -99,7 +99,7 @@ namespace tktk
         return m_motion->getEndFrameNo(handle);
     }
 
-    void MeshResource::updateMotion(size_t skeletonHandle, size_t curMotionHandle, size_t preMotionHandle, size_t curFrame, size_t preFrame, float amount)
+    void MeshResource::updateMotion(size_t skeletonHandle, size_t curMotionHandle, size_t preMotionHandle, float curFrame, float preFrame, float amount)
     {
         m_skeleton->transform(skeletonHandle, m_motion->calculateBoneTransformMatrices(curMotionHandle, preMotionHandle, curFrame, preFrame, amount));
     }
